+++ lib/src/utils/code_part_widget.dart
@@ -1,17 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:international_phone_text_field/international_phone_text_field.dart';
 import 'package:international_phone_text_field/src/controller/phone_controller_bloc.dart';
 
 class CodePartWidget extends StatelessWidget {
   const CodePartWidget({
     super.key,
-    required this.widget,
     required this.codeController,
     required FocusNode codeFocusNode,
     required this.controllerBloc,
+    required this.style,
+    required this.cursorColor,
   }) : _codeFocusNode = codeFocusNode;
 
-  final InternationalPhoneTextField widget;
+  final TextStyle style;
+  final Color cursorColor;
   final TextEditingController codeController;
   final FocusNode _codeFocusNode;
   final PhoneControllerBloc controllerBloc;
@@ -25,7 +26,7 @@ class CodePartWidget extends StatelessWidget {
         mainAxisAlignment: MainAxisAlignment.center,
         crossAxisAlignment: CrossAxisAlignment.center,
         children: [
-          Text("+", style: widget.style),
+          Text("+", style: style),
           Flexible(
             fit: FlexFit.loose,
             child: IntrinsicWidth(
@@ -35,7 +36,7 @@ class CodePartWidget extends StatelessWidget {
                 maxLines: 1,
                 maxLength: 4,
                 focusNode: _codeFocusNode,
-                style: widget.style,
+                style: style,
                 decoration: InputDecoration(
                   counterText: "",
                   border: InputBorder.none,
@@ -44,7 +45,7 @@ class CodePartWidget extends StatelessWidget {
                   focusedBorder: InputBorder.none,
                   errorBorder: InputBorder.none,
                 ),
-                cursorColor: widget.cursorColor,
+                cursorColor: cursorColor,
                 onChanged: (String text) {
                   controllerBloc.add(FindCountryCode(code: text));
                 },