+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 ## 0.0.1+1
 
 - Update README.md
+
+## 0.0.1+2
+
+- Added example file to the package
+- Updated all packages to the latest version
+- Some minor changes in the code